refactored app to be outside of the session manager class
