+++ package.json
@@ -5,9 +5,8 @@
   "main": "index.js",
   "scripts": {
     "init": "npm install && git submodule init && git submodule update && npm run init --prefix ./front/",
-    "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc",
-    "start": "ts-node ./src/index.ts",
+    "start": "node ./dist/index.js",
     "run": "npm run build && npm run start",
     "random_secret": "node ./dev-scripts/generate_secret_token.js"
   },
