+++ tsconfig.json
@@ -1,5 +1,9 @@
 {
   "compilerOptions": {
+      "outDir": "./dist",
+      "rootDir": "./src",
+      "module": "commonjs",
+
     /* Visit https://aka.ms/tsconfig to read more about this file */
 
     /* Projects */
@@ -106,5 +110,7 @@
     /* Completeness */
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
-  }
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules"]
 }
